:truck: Move Logger to utility module and separate functions into separate module
