+++ data_safe_haven/administration/users/user_handler.py
@@ -6,7 +6,8 @@ from typing import Any, Dict, List, Sequence
 from data_safe_haven.config import Config
 from data_safe_haven.exceptions import DataSafeHavenUserHandlingException
 from data_safe_haven.external.api import GraphApi
-from data_safe_haven.mixins import AzureMixin, Logger
+from data_safe_haven.mixins import AzureMixin
+from data_safe_haven.utility import Logger
 from .active_directory_users import ActiveDirectoryUsers
 from .azure_ad_users import AzureADUsers
 from .guacamole_users import GuacamoleUsers
