+++ data_safe_haven/commands/deploy_shm_command.py
@@ -16,10 +16,10 @@ from data_safe_haven.exceptions import (
     DataSafeHavenInputException,
 )
 from data_safe_haven.external.api import GraphApi
-from data_safe_haven.helpers import password
-from data_safe_haven.mixins import Logger
+from data_safe_haven.functions import password
 from data_safe_haven.provisioning import SHMProvisioningManager
 from data_safe_haven.pulumi import PulumiStack
+from data_safe_haven.utility import Logger
 
 
 class DeploySHMCommand(Command):  # type: ignore
