+++ data_safe_haven/commands/deploy_sre_command.py
@@ -14,11 +14,10 @@ from data_safe_haven.exceptions import (
     DataSafeHavenInputException,
 )
 from data_safe_haven.external.api import AzureApi, GraphApi
-from data_safe_haven.helpers import alphanumeric, password
-from data_safe_haven.mixins import Logger
+from data_safe_haven.functions import alphanumeric, password
 from data_safe_haven.provisioning import SREProvisioningManager
 from data_safe_haven.pulumi import PulumiStack
-
+from data_safe_haven.utility import Logger
 
 class DeploySRECommand(Command):  # type: ignore
     """
