+++ data_safe_haven/commands/teardown_backend_command.py
@@ -12,7 +12,7 @@ from data_safe_haven.exceptions import (
     DataSafeHavenException,
     DataSafeHavenInputException,
 )
-from data_safe_haven.mixins import Logger
+from data_safe_haven.utility import Logger
 
 
 class TeardownBackendCommand(Command):  # type: ignore
