+++ data_safe_haven/commands/teardown_shm_command.py
@@ -11,8 +11,8 @@ from data_safe_haven.exceptions import (
     DataSafeHavenException,
     DataSafeHavenInputException,
 )
-from data_safe_haven.mixins import Logger
 from data_safe_haven.pulumi import PulumiStack
+from data_safe_haven.utility import Logger
 
 
 class TeardownSHMCommand(Command):  # type: ignore
