+++ data_safe_haven/config/config.py
@@ -14,8 +14,9 @@ from azure.storage.blob import BlobServiceClient
 from data_safe_haven import __version__
 from data_safe_haven.exceptions import DataSafeHavenAzureException
 from data_safe_haven.external.api import AzureApi
-from data_safe_haven.helpers import alphanumeric
-from data_safe_haven.mixins import AzureMixin, Logger
+from data_safe_haven.functions import alphanumeric
+from data_safe_haven.mixins import AzureMixin
+from data_safe_haven.utility import Logger
 
 
 class Config(AzureMixin):
