+++ data_safe_haven/mixins/__init__.py
@@ -1,7 +1,5 @@
 from .azure_mixin import AzureMixin
-from .logging_mixin import Logger
 
 __all__ = [
     "AzureMixin",
-    "Logger",
 ]
