+++ data_safe_haven/pulumi/components/shm_data.py
@@ -8,7 +8,7 @@ from pulumi_azure_native import keyvault, resources, storage
 
 # Local imports
 from data_safe_haven.external.interface import AzureIPv4Range
-from data_safe_haven.helpers.functions import (
+from data_safe_haven.functions import (
     alphanumeric,
     replace_separators,
     truncate_tokens,
