+++ data_safe_haven/pulumi/components/shm_update_servers.py
@@ -8,7 +8,7 @@ from pulumi import ComponentResource, Input, Output, ResourceOptions
 from pulumi_azure_native import network
 
 # Local imports
-from data_safe_haven.helpers import b64encode
+from data_safe_haven.functions import b64encode
 from data_safe_haven.pulumi.common.transformations import (
     get_available_ips_from_subnet,
     get_name_from_subnet,
