+++ data_safe_haven/pulumi/components/sre_gitea_server.py
@@ -7,7 +7,7 @@ from pulumi import ComponentResource, Input, Output, ResourceOptions
 from pulumi_azure_native import containerinstance, dbforpostgresql, network, storage
 
 # Local imports
-from data_safe_haven.helpers import FileReader
+from data_safe_haven.utility import FileReader
 from ..common.transformations import (
     get_ip_address_from_container_group,
     get_ip_addresses_from_private_endpoint,
