+++ data_safe_haven/utility/__init__.py
@@ -1,0 +1,9 @@
+from .file_reader import FileReader
+from .logger import Logger
+from .types import PathType
+
+__all__ = [
+    "FileReader",
+    "Logger",
+    "PathType",
+]
