+++ data_safe_haven/utility/types.py
